update hook test
